Replaces GameOverModal with MultiplayerGameOverModal
This commit replaces the GameOverModal component with MultiplayerGameOverModal in both the PassAndPlay and Play components.

This change likely aims to unify the game over modal across different game modes (pass and play, and online play) to provide a consistent user experience.
